Move Elasticsearch code independent of Hail to shared folder
